saving json file works, first test on cell 0, testing load to create a cell list

+++ The_LabyRun/Assets/Script/Map/map_save/saver.cs
@@ -10,20 +10,29 @@ namespace sav
 
         public void Saving(string txt, string fichier, bool overwrite = true)
         {
-            if (File.Exists(Application.dataPath+fichier))
+            if (File.Exists(Application.dataPath+fichier) && overwrite)
             {
-                if (overwrite)
-                {
+                
                     File.WriteAllText(Application.dataPath+fichier, txt);
-                }
-                else
-                {
-                    File.AppendAllText(Application.dataPath+fichier, txt);
-                }
+                
             }
             else
             {
                 File.AppendAllText(Application.dataPath+fichier, txt);
+            }
+        }
+
+        public void Savinglines(string[] txt, string fichier, bool overwrite = true)
+        {
+            if (File.Exists(Application.dataPath + fichier) && overwrite)
+            {
+
+                File.WriteAllLines(Application.dataPath + fichier, txt);
+
+            }
+            else
+            {
+                File.AppendAllLines(Application.dataPath + fichier, txt);
             }
         }
 
@@ -42,6 +51,20 @@ namespace sav
             return ext;
         }
 
+        public string[] Loadlines(string fichier)
+        {
+            string[] ext;
+            if (File.Exists(Application.dataPath + fichier))
+            {
+                ext = File.ReadAllLines(Application.dataPath + fichier);
+            }
+            else
+            {
+                throw new System.Exception("File not found, wrong path or deleted file");
+            }
+
+            return ext;
+        }
 
     }
 }
